// src/screens/RegistrationScreen/types.ts
export type UserDataObject = {
  photo: string;
  google_id: string;
  prefix: string;
  first_name: string;
  middal_name: string;
  last_name: string;
  email: string;
  mobile_number: number;
  designation_id: number;
  job_status: string;
  office_address: string;
  office_district_id: number;
  native_district_id: number;
  office_id_photo: string;
  leaving_certificate_photo: string;
};